Convert to typescript :)
